#Added feature to make copy

+++ app/src/main/java/com/mintanable/notethepad/feature_note/domain/model/Note.kt
@@ -15,7 +15,7 @@ import java.lang.Exception
     ]
 )
 data class Note(
-    @PrimaryKey(autoGenerate = true) val id: Int? = null,
+    @PrimaryKey(autoGenerate = true) val id: Long? = null,
     val title: String,
     val content: String,
     val timestamp: Long,
