+++ app/src/main/java/com/mintanable/notethepad/feature_note/domain/use_case/GetNote.kt
@@ -7,7 +7,7 @@ class GetNote(
     private val repository: NoteRepository
 ) {
 
-    suspend operator fun invoke(id:Int): Note?{
+    suspend operator fun invoke(id:Long): Note?{
         return repository.getNoteById(id)
     }
 }
